3.3 authorize the profile, redirect if not logged in
This is a new one, and intros some new pieces of Angular. In the README
there are links associated with Step 3 and 3.3 you should read for more
info.

The main explanation of what's happening is in the code comments itself;
we are essentially adding a state change event listener to our app after
registering our states that checks a custom property we add to any
states we want to authorize.

One big thing to notice here is we are breaking the John Papa rule of
one component per file. This is because I think it's worth it to isolate
all of the routing logic, including authorizing routes, in one place.