4.1 (second) add profile controller; fill inputs
Here we do the standard: create a controller, declare it on the state
in the app.routes.js, link it in index.html, and bind it to the form
with ng-model. Alles Gute.

However, we pre-populate the form data with the current user data
stored in the token by adding a new method .currentUser() to our auth
service, which decodes and cleans up the token data. We reuse the
method throughout the template instead of binding its value, since
we want the values to either be static or changeable due to
circumstances (the header vs the input fields vs the token itself,
eg).