4.2a (first) server/api update a user
There is a route (PUT /users/me) that allows an authenticated user to
send an update request. The handler is a little odd: it edits the
found user and saves it, instead of updating it. The reasons are
explained in the comments. When successful, due to how Mongoose works,
it only returns a 200 OK, not a copy of the updated user.

Note that the update uses the decoded id, not email, to find the
authenticated user.