4.2b (first) server/api add token refresh
Instead of having to re-create a brand new token (in which you need to
send authentication "credentials," ie your password, along with the
request), this route generates a new token for the authenticated user.

This means, if you've updated the user, that you can get the updated
data in a new token without having to sign out and sign in!