REFACTOR 2: add http interceptor for Content-Type
First, we tell $http to always set the header for any request to
Content-Type: application/json. To be honest, this is a trivial little
thing, and not always something you should do: it's more because I've
set the default server to enforce this rule for all requests in order
to habituate us to thinking about headers.

It's an example, really. The more important interceptors come next!