REFACTOR 2: add http interceptor for Authorization
Does what we just did with the Content-Type header commits all at once,
moving the addition of an Authorization header to any request made by
$http if there is a token present.

Note: we are using the tokenService instead of the authService here, bc
of a "cyclical dependency": since authService injects $http, and $http
injects $httpProvider, if we inject authService into $httpProvider we
would go into an injection loop until the sun exploded. Let's avoid
that and just use the tokenService instead.